+++ lib/Controller/API/V2/AbstractDownloadController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace API\App;
+namespace API\V2;
 
+use API\App\AbstractStatefulKleinController;
 use AuthCookie;
 use Bootstrap;
 use CookieManager;
